added file animals.js. moved animals class into the file and exported it. imported the class to index. kevin.jonsson@medieinstitutet.se
